Initial import of kyua-cli
The kyua-cli package provides the main CLI tool for the Kyua testing
framework.  The main purpose of this tool is to allow executing test
cases, collecting their results and allowing inspection of these
results data a posteriori.

The current code for this import implements a very simple 'kyua' binary
that is already able to run its own test suite.  This code also provides
many helper utility functions and classes that may need to be split out
into a separate package later on.

This change was r2 in Subversion.