Add tests to validate the currently-implemented test case isolation features
These tests include validation of the environment cleaning, the creation of
a new process group, the resetting of the umask and the execution in a new,
temporary work directory.  There are still things to be done in this area as
detailed by the TODOs in the code.

While doing so, fix a couple of issues.  Namely, correctly set the value of
the sane HOME variable to an absolute path and nicely report errors during
the isolation preparation.

This change was r7 in Subversion.