Add timeout support to the wait of children processes
Subprocesses (as well as their subprocesses if they are in the same process
group) are forcibly killed when the given timeout expires.  This will be used
to execute the test cases in a controlled manner (to not allow a single test
to stall the execution of others).

While here, move the create_file auxiliary function from
utils/fs/operations_test to the utils/test_utils module so that it can be
reused in our new tests.

This change was r14 in Subversion.