Move the timedelta class out of the signals module into a datetime module
Create a new utils/datetime module to hold the "delta" class, which was
previously found in utils/signals/timer.  It does not make sense to have
to pull in the "heavy" timer definition just to get the ability to
represent time deltas.

(This module should really be called "time" at this point but that
collides with time.h.  Just name it "datetime" and it will be more
future-proof... and Pythonic!)

This change was r15 in Subversion.