Detect the last signal number and ensure all signals are reset in test cases
The test case that was validating whether all signals are reset or not during
the execution of test cases was broken: it misdetected the fact that SIGUSR2
was not being reset.  After fixing the test case, I realized that last_signo
was too low for the current platform, so here comes the real fix.

This change was r16 in Subversion.