Build and runtime fixes for NetBSD
Rename some variables and namespaces to prevent collisions with more global
names.  Dunno why clang in Mac OS X did not pick these up.

Also add an explicit flush of std::cout in a test so that the ordering of
the stdout and stderr output is deterministic.  This fixes a test at run
time.

This change was r18 in Subversion.