Refactor the test_case representation and implement metadata validation
This change reimplements the whole test_case module to better represent
test cases.  Properties that have a special meaning internally are
explicitly represented as class fields using appropriate types;
user-specific properties are kept as a free-form string pair in a
dictionary (which is what we used to do for all properties before).

At the same time, this implements proper validation of the test case
metadata properties (alongside with extensive tests).  Note that, at
this point, most of these properties are still ignored by the runtime
engine.

This change was r19 in Subversion.