Implement enforcement of the require.{arch,machine} properties
Add support to check that the architecture/platform requested by the
test case matches the ones in which the test case is currently running.

This change adds a placeholder for a global configuration object to
hold overrides for the default architecture and platform values detected
by the configure script.  This module will be improved at a later stage
support more configuration tunables and to allow reading custom values
from a configuration file; in particular, we must also bundle the
properties map object that we now pass around as configuration variables.

Update issue 7

This change was r25 in Subversion.