Parametrize dependencies to internal libraries
For every internal library, define the <libname>_CFLAGS and <libname>_LIBS
variables that contain the necessary C compiler and linker flags to link
against such internal libraries.  This will make it simpler to add
dependencies on external components (e.g. Lua or SQLite) later on.

Be aware that this is technically wrong: for our purposes, <libname>_CFLAGS
should really be named <libname>_CXXFLAGS and/or <libname>_CPPFLAGS.
Similarly, some of the flags that end up in <libname>_LIBS should really be
in <libname>_LDFLAGS.  However, I am following this simplified (but, again,
invalid!) naming scheme to match the variables defined by pkg-config when it
detects libraries.  Consistency will make things clearer.

This change was r37 in Subversion.