Rename the test suite configuration code and files to kyuafile
The engine/suite_config module is now called engine/kyuafile and the
kyua.suite files are now named Kyuafile.

This is to follow the standard practice of other build tools (make, atf
and doxygen to name a few) in which the name of their control files is
capitalized and carries the 'file' suffix.  Doing this change before I
start working on the new code to parse test suite descriptions.

Note that the old suite_config module name carried the implied meaning of
"test suite" whereas Kyuafile does not.  I am not yet sure if doing this
rename is good or bad, but in the worst case we can just reinstantiate a
test-suite module later (e.g. keep kyuafile purely to parse files and add
another module to represent a test suite in a conceptual manner).  That
said, drawing the line between a Kyuafile and a test suite is hard, very
hard... to the point where they may actually be representing the same
concept.

This change was r41 in Subversion.