Implement a mechanism to create Lua modules from C++
A module in this context is just a table with a collection of native
functions in it.  We will use this to expose some internal Kyua functions
to the Lua scripts.

This change comes with a bunch of new wrapping methods in the C++
Lua state required to make the create_module function work.  The wrapping
of C++ methods is particularly tricky (and, indeed, ugly).

This change was r42 in Subversion.