Add Lua bindings for utils::fs
These bindings provide a minimum set of functionality to manipulate path
names but, supposedly, these are the only functions required to implement
the 'include' functionality of Kyuafiles.  (If there is anything missing
it will be implemented later.)

This change also comes with the addition of a new lua::state::is_function()
method.

This change was r44 in Subversion.