Ensure call_cxx_function_from_c does not leak C++ objects
There is the possibility that running luaL_error from within an exception
catch clause can leak C++ exception objects (and, therefore, memory) because
Lua internally uses longjmp to get out of the function.

To workaround this possibility, store the error message in a local variable,
exit any C++ scope, and raise the error.  (One of the very few things that
longjmp is guaranteed to clear is the stack.)

This change was r45 in Subversion.