Fix crashes when calling the native methods with incorrect types
Our Lua binding functions cannot blindly assume that the parameters on the
Lua stack are of the correct type, so issuing a state::to_* without prior
checking with a corresponding state::is_* is invalid.

Therefore, properly validate the types of all the input parameters to our
fs binding functions.  Also add tests to validate correct results (and, in
particular, no crashes due to assertion failures).

(Note: An alternative would be to change state::to_* to raise exceptions
instead of precondition failuers in case of type mismatches.  This seems
like a reasonable thing to do because we can be dealing with untrusted
input.  I may do this later, but conceptually, I like forcing the need to
call is_* before to_* as this gives more control on how errors are
reported.)

This change was r46 in Subversion.