Convert the definition of Kyuafiles to Lua
This change adds the processing of Lua-based Kyuafiles.  All this new code,
which involves some host C++ code as well as a few supporting Lua libraries,
goes into a new engine::user_files namespace.  This namespace will later
contain routines to process other user-provided files (such as configuration
files) and is why it has such a generic name.

There is still quite a bit of cleanup work to be done in the
engine::user_files::kyuafile module, which is basically what the old
engine::kyuafile module contained.  However, this change is already so big
and fragile that I am submitting it now that it is in a reasonable state.
The code will be polished later and improved upon iteratively.

As a side-effect of this change, all the Kyuafiles in the source tree have
been migrated to the new syntax.

This change was r49 in Subversion.