Ensure getopt_long does not print errors by itself
Explicitly set opterr to 0 while processing command line options so that,
in the case of errors, getopt_long does not print them.  In principle
this is what the ':' character at the beginning of the string (but after
the '+' character) should be doing, but it does not seem to work in
NetBSD.  It could be a NetBSD bug though -- need to check what the GNU
implementation does.

While doing this, add an extra test to ensure that getopt_long does not
recognize the leading '+:' sequence as a '+' short option with a required
argument.

This change was r50 in Subversion.