Fix detection of the name of the option that caused an error
The code was using argv[optind - 1] to find out the name of the option that
caused a getopt error (i.e. invalid option or missing argument).  This is
completely broken and, in particular, it fails miserably when providing
short options in blocks (e.g. -abc).  Fix this by looking up the name
accordingly, which prevents some assertions from triggering.

While doing this, also detect a broken getopt_long in Ubuntu 10.04.1 and
report the inconsistencies caused by it as expected failures (instead of
saying that the test cases failed because it is not really our fault!).

This change was r54 in Subversion.