Make functions in utils::lua that open files report missing files accordingly
Modify lua::state::load_file and lua::do_file to report missing files with
a special exception instead of bundling the error in more generic error types.
We may need to be able to detect whether a file exists or not as, in some
cases, such condition is not fatal (think about an optional configuration
file!).

This change was r59 in Subversion.