Initial support for Lua-based configuration files
This change adds very rudimentary support to load and process Lua-based
configuration files for the Kyua runtime engine.

This is still very incomplete but already allows us to inject
configuration variables from the outside; e.g. the value of
"unprivileged-user" has been always hardcoded in the source tree for a
long time already and it was impossible to change its value.

Some important things are yet to be done.  For example:

- Nuke the user_config strings map from the runtime engine code.
  The test-suite configuration variables should be defined in the
  configuration file as well, although I am not yet sure what the best
  way to do this is yet.

- Properly locate the configuration file in sysconfdir instead of
  hardcoding the name of the default file.  (But the cli code is still
  hackish, hence why I have not done it properly in this change.)

This change was r60 in Subversion.