Define per-Kyuafile test suites
Allow each Kyuafile to define the test suite to which it belongs.  This
information is needed by the configuration system to know which variables
to apply to every test.

This approach to test suites is inherited from ATF.  I don't think it is
a very good design, but we need it for now to remain compatible.  We can
revisit the approach later.

This change was r64 in Subversion.