Propagate per-test-suite variables from configuration files to test cases
The previous changes added support to the configuration files to define
per-test-suite configuration variables (r63) and changed the Kyuafiles to
be able to specify which test suite they belong to (r64).

This change brings the previous two changes together by making it possible
to pass the per-test-suite configuration variables defined in the
configuration files to the test programs that will make use of them.

This kills the user_config abomination added a while ago to represent the
configuration variables passed to test programs.  The results are not
that much nicer at this point, but I really want to get things out of the
way before doing any refactorings.

This change was r65 in Subversion.