Kill the lua::get_array_as_strings utility function
This was added to support the implementation of the Kyuafile processing
but it is not necessary any more.  Furthermore, the level of abstraction
was incorrect since Kyuafile needed paths instead of strings.  Let's
just leave the task of Lua data processing to any module that needs it.

This change was r67 in Subversion.