Report path concatenation errors as exceptions, not assertions
As the TODO comment in path::operator/ prophesied, using assertions to
detect invalid arguments to paths concatenations can lead to crashes
because we very easily end up swallowing user-provided paths.  Change
this to report errors by means of exceptions.

The particular case that lead to this error is:
1. Kyuafile includes dir/Kyuafile.
2. dir/Kyuafile includes /Kyuafile.
3. When 1. recovers control to merge the results of 2., it tries to join
   the paths dir/ and /Kyuafile and crashes.
I need to add better control in the kyuafile module for this kind of
problem, but that is separate.

This change was r68 in Subversion.