Generalize the definition of the --config and --kyuafile flags
This change adds a new cli/common module that contains standard definitions
for the --config and --kyuafile options and some helper code to deal with
their values and the files they refer to.

As part of this change, the default configuration files are now properly
looked up in the user's home directory and in the sysconfdir.

Fixes issue 10.

This change was r70 in Subversion.