Detect missing test programs when processing a Kyuafile
Bail out early when processing a Kyuafile that references missing test
programs.

In the past, some people mentioned that erroring out when an Atffile
refers to a test program that does not exist, atf should just continue
running and report the test program as failed.  To me, the fact that
a configuration file refers to a non-existent test program is a serious
error in the definition of the test suite and it should be treated as
such.  Hence why I am following the same approach in the Kyuafiles and
just report the failure loudly.

Actually, should this have been in place, I would have caught earlier
the fact that config_test was moved from engine/ to engine/user_files/
but the Kyuafile was not updated accordingly (see r71).

This change was r72 in Subversion.