Add a placeholder info file for the manual to the Kyua command-line tool
This change adds a minimal info file to act as a placeholder for further
documentation of the kyua-cli package.  At the moment, I am adding a tiny
overview copy-pasted from the web page and a chapter with the license
information.

Update issue 12
The placeholder has been added.  I still need to add a minimal manual page
that refers to the info documentation so that 'man kyua' works.

This change was r73 in Subversion.