Prevent static initialization of subcommand classes
We can't control the order in which different translation units are
initialized: statically constructing base_command objects is bound to
failure because they often have dependencies on other objects.

This fixes an immediate crash in Mac OS X with clang++.  This should have
been expected though, as the previous behavior was completely broken.

This change also gets us rid of the "all_commands" ugly module, which
was a singleton that maintained the list of commands for the CLI.  This
list is now injected as part of the call to main, which makes testing
slightly easier.

This change was r75 in Subversion.