Add a timestamp class to be able to query and manipulate the current time
This class is very minimal at this point and only provides a mechanism to
capture the current time and format it to any random string.  This will
be used to add timestamps to log entries (and later, for example, to
calculate how long it takes to run the tests).

I'm using std::tr1::shared_ptr in this change and this is the first use
of tr1 in the codebase... but this makes the implementation of the
timestamp class much simpler.  Let's see how long this lasts before the
lack of tr1 bites us!

Update issue 14
We can now add timestamps to every log entry

This change was r80 in Subversion.