Add a basic logging library
The logging library provides a simple mechanism to generate log files of
the execution of a binary.  This will be used by the Kyua cli library to
trace internal actions so that it is easier to perform postmortem analysis.

Update issue 14
With this, we can now update the entry point of the program to initialize
the logging library and then sprinkle calls to the logging macros across
the sources.

This change was r82 in Subversion.