Implement the --logfile option
The --logfile option is an application-wide option to specify where to log
activity messages to.  This option always has a default value to ensure that
messages are logged so that they come in useful for postmortem analysis.

Update issue 14
Still to do: fix 'help' to print global options, add some messages in random
modules and find a way to print the path to the log file in case of crashes.

This change was r85 in Subversion.