Add the source file and line number to log messages
Log messages are more useful when they point to the source file and the
line number that generated them.  Otherwise they are too vague.  Make
this happen.

Update issue 14
This is just a tiny improvement over the previous code that should have
gone in since the beginning.

This change was r86 in Subversion.