Make 'help' show information about program-wide options
The 'help' command now prints details about options that are applicable
to all commands both in the general program help and in the subcommand
help.

To achieve this, I had to do a little bit of code shuffling.  Injecting
the commands to the cli::main() function from ::main() was just a bad
idea because that meant I had to do instantiate the options in the same
place to pass them to cmd_help.

I have now moved the instantiation of commands into cli::main() (as it
is how it should be done).  This allows us to pass the global program
options to cmd_help.

Update issue 14
The need to print global options was exposed by the addition of the
--logfile flag.

This change was r89 in Subversion.