Fix processing of options with default values
For our use cases, a flag can be optional but the argument it takes cannot.
Something like 'kyua --logfile /tmp/a help' (note the lack of = between
the flag name and the path) was broken because we implemented optional
arguments to flags by mistake.  Fix this.

We do not actually want to support optional flags with optional arguments;
this was just a misunderstanding from my side regarding the
optional_argument and required_argument features of getopt_long.

This change was r90 in Subversion.