Ensure that subprocesses of a test case are killed on exit
If a subprocess spawns children processes, we must ensure that they get
killed when the test case terminates, or otherwise they will clog the
system.  This problem has been found when running the file system tests
in NetBSD (src/tests/fs/vfs/ in particular).

To do so, move the logic of setting up process groups into the process
library and forcibly terminate all the processes in the group when we
wait for the termination of the group leader.  While doing this, add
tests to validate these semantics from the test case runner layer.

Update issue 16
We need to implement integration tests and add this particular case to
them.  The unit tests we currently have are very good now but they will
be brittle during future refactorings.

This change was r95 in Subversion.