Implement preliminary test case filtering for 'list'
This change adds new functionality to allow the end user to specify the
test cases to be listed in a fine-grained manner.  The same functionality
will later be used in the 'test' subcommand.

There are many rough edges in this change and I suspect the filtering
itself will need to change a bit.  But it's good enough for now to let
several extra tests pass... and the change is already big enough to be
sent before I do further modifications.

In particular: this deprecates some code that must go away and this does
not address other shortcomings of the 'list' command (e.g. --verbose).

This change was r103 in Subversion.