Implement preliminary test case filtering for 'test'
Mimic the implementation of the test case filtering previously added
to 'list' and add it to 'test'.  This is mostly to get rid of obsoleted
code (the old_load_kyuafile function).

As a result, we can also get rid of the "runner hooks", which arguably
were pretty ugly.  The test case filtering happens now at the UI level,
so we do all the test program processing from the 'test' command itself.

This comes with no new tests.  I am postponing their addition until the
'list' command is clean enough -- to the point where its tests can be
easily ported to 'test'.

This change was r104 in Subversion.