Force test program names to be relative to their test suite
The location of the test suite definition (aka the location of the first
Kyuafile processed) is now considered to be the root directory of the test
suite.  All further paths to test programs must be relative to this root.
We don't do great efforts to enforce this (yet?), but we do basic sanity
checks.

The point of this change is to ensure consistency in the processing of the
test case filters.  If someone does:

    kyua test -k /usr/tests/Kyuafile foo/bar_test

we want foo/bar_test to be /usr/tests/foo/bar_test and we don't want the
user to have to type the whole path in the filter (because that's not
conceptually accurate!).

This change was r109 in Subversion.