Flush the log file after setting the persistency settings
Ensure the log file is flushed after we set the persistency so that tests
that want to inspect the log right after really see all the data.  I am
not sure why the tests have suddenly started failing, but this is obviously
a bug.

This change was r114 in Subversion.