Catch and print runtime errors
Make the main function catch any std::runtime_error exceptions and print
them as errors.  The original idea was to make any subcommand do this on
its own, but I see no value now in proceeding that way.

While doing this, adjust the 'list' integration tests to properly validate
error messages.  The error messages are not very pretty now, but they are
good enough.

This change was r116 in Subversion.