Show the path to the log file upon a crash
The intention is to allow end users who suffer a program failure to be
able to dig down more details on what happened and to provide better
bug reports.

Update issue 14
This had been planned for a while as part of the implementation of the
logging module but never happened.  Do so now.

This change was r121 in Subversion.