Implement integration tests for the 'test' command
Most of the new tests are a copy/paste from the tests of 'list', with the
appropriate adjustments to make them work for 'test'.

Some of these tests are marked as expected failures due to some shortcomings
in the current implementation of 'test'.  These will be addressed separately.

Also, it's hard to decide where to draw the line in what to test here. So far
I have kept the tests restricted to the behavior of the 'test' command, but
we probably want to go further and ensure that the runtime engine behaves
correctly when running any kind of test programs (isolation, etc.).  Maybe
these belong somewhere else though...

This change was r124 in Subversion.