Add support to get multi-option values
A "multi-option" is an command-line option that can be specified multiple
times and all the provided values have to be aggregated in some form.
This is required to implement the behavior of the '--variable' flag for
the 'test' command, which will allow the user to override configuration
variables of test suites.

This change was r125 in Subversion.