Implement the config::apply_overrides method
The apply_overrides method for configuration objects provides a way to
process user-specified configuration settings and applies them to an
existing configuration object.  This method can take the values of a
property_option flag, and thus will be used to implement the --variable
flag.

As mentioned in a todo comment, this is overly complicated because we
are implementing custom parsing for property assignments.  Maybe we
could reuse Lua in the future to keep consistency with the handling of
the properties in configuration files and avoid some code.

This change was r127 in Subversion.