Suppress "running outside of atf-run" warning messages in test programs
Since atf-0.14, test programs print a warning message when they are not
being run from atf-run.  This is to prevent user confusion: the
implementation of test case isolation happens at the test runner level,
not at the test case level, and thus the test case may report invalid
results.

Do the same in kyua: define the necessary magic environment variable to
suppress these warning messages.  Unfortunately, there is no way for us
to test the validity of this just yet because kyua does not provide any
way to inspect the output of test programs.

While doing this, also document the test program interface change.

This change was r132 in Subversion.