Implement the fs::unmount function
As the name implies, fs::unmount is a portable function to unmount a file
system.  This change comes with some simple tests, although they are more
complex than they should because mouting a tmpfs portably is hard.

Update issue 17
This functionality will soon be used in the work directory cleanup code
to ensure stale mount points left by the test case are properly cleaned.

This change was r134 in Subversion.