Implement unmounting of file systems from within work directories
This change reworks the way the fs::cleanup function works altogether
and implements support to unmount file systems from within work
directories.

This implementation mimics the code in atf-run but should be significantly
more robust.  First of all, it deals with the case of a paticular mount
point having multiple file systems stacked on top.  Secondly, error
checking is tightly controlled to ensure undesired errors do not propagate
up to the caller and kill the runtime engine.  And lastly, the code
includes several logging calls to aid in debugging later on.

Update issue 17
Testing for this functionality is quite tricky and unportable.  While I
think the new code should do the right thing and cover the problems we
discovered before in atf-run, I'm not sure it does.  I still have to try
this on the NetBSD tests to see how things go.

This change was r135 in Subversion.