Fix test execution in OS X
Paths to temporary files in OS X can get very long, and we were creating a
configuration file in one of these temporary directories.  For some reason,
Lua truncates these paths in error messages when they are too long, and
this was breaking a test.

Instead of doing exact messsage matching, use a regular expression to be
more relaxed on the expected output of the application in this particular
case.

This change was r137 in Subversion.