Make the wait_killpg tests tests more robust
Sending a kill signal to a process does not mean that the process will
be gone from the system immediately.  I.e. delivery of signals is not
synchronous.

Fix a couple of tests that were assuming this by attempting the wait
for the subprocesses death a few times.  This resolves a few random
failures in OS X.

This change was r138 in Subversion.