Implement signal handling and cleanup of active test cases
Make the test case runner capture some common termination signals and
force it to terminate the active test case before exiting.

This comes with some assorted changes to the process module in order
to support the implementation of this feature.  In particular, the
eager child*::wait functions that would retry on EINTR made the
implementation of this controlled exit hard, so they had to be modified.

Fixes issue 4
This seems to work fine in all cases I tried and in the automated tests
for this new feature.  That said, I am sure there is some corner case
that is not covered and will result in turds being left behind after a
user cancels the execution.  However, because I have not been able to
uncover these cases at the moment, I see no value in keeping the ticket
open.  So... one less thing to do before 0.1!

This change was r139 in Subversion.