Add wrappers to manipulate userdata objects, C closures and metatables
These functions are all required to implement iterators from native
code.

Update issue 9
Iterators will be used to implement recursion in Kyuafiles.  For example,
something along the lines of: "for file in files('directory')".  Haven't
decided yet how to expose this in Kyuafiles though, but this change is
self-contained enough to be submitted now.

This change was r142 in Subversion.