Implement the exists and files functions
exists is a simple wrapper over fs::exists to check for the existence of
a given file.

files is an iterator to scan the contents of a directory.

Update issue 9
These two functions are going to be used to implement recursion from the
top-level directory of a tests hierarchy.

This change was r143 in Subversion.