Add a sample Kyuafile for the top-level directory of a test suite
This sample Kyuafile, named Kyuafile.top in the samples directory, scans the
directory it is in for any existing */Kyuafile files and sources them.

This file is intended to be installed in the top-level directory of a test
suite (e.g. /usr/tests) to provide automatic recursion into the tests of any
installed packages (without having the packages touch the top-level Kyuafile
to register themselves!).

Fixes issue 9.

This change was r144 in Subversion.