Improve error messages when a test case fails to write a results file
When a test case does not write a results file for whatever reason, it is
better to report the error condition of the test rather than to say that
a results file (which points inside an ephemeral temporary directory)
cannot be opened.

Update issue 13
The improvement of error messages when the test case terminates prematurely
addresses part of this issue.

This change was r145 in Subversion.