Prevent a double-unprogramming of interrupts
The previous code was attempting to unprogram interrupts right after they
had been unprogrammed in the rare condition were the signal was received
after the test case finished its cleanup.

Do this by moving the unprogramming out of the try-block that also attempts
to do the unprogramming in error conditions.

This change was r146 in Subversion.