Expose exec(3) errors from process::exec
The process::exec function used to print any errors in exec(3) on its
own and forcibly terminated the subprocess to prevent the execution
from returning to the caller at all.  This is suboptimal because it
forbids the caller to inspect exec errors.  Therefore, make the exec
wrapper raise an exception when the call to exec(3) fails, just like
any of the other process calls.

As a side-effect of this, update the fork() functions to ensure that
neither exceptions nor normal returns propagate back to the caller in
the case of the subprocess.  This is more robust than what we were
doing before.

Update issue 13
This change is required to permit the engine to report nicer error
messages when there are problems executing test program binaries.

This change was r147 in Subversion.