Clarify error messages when test programs fail during listing
If test programs cannot be executed for whatever reason or if they exit
unexpectedly while printing their test case list, report such cases
accordingly.  The previous code was confusingly reporting parse errors
in the output (empty on exec failures, for example), which are not
helpful at all.

Update issue 13
This change addresses part of the problems reported in this ticket.
It clarifies a potential source of confusion when test programs are
really broken and fail even in their 'list' operation.

This change was r148 in Subversion.