Add tests to ensure error messages are valid on missing test programs
This change is only to ensure that Kyua prints useful error messages
when Kyuafiles refer to missing test programs.  Note that this renames
previously-existing test cases that had inappropriate names and
conflicted with the new tests.

Update issue 13
Ensures some error conditions are reported in a useful manner.

This change was r149 in Subversion.