Install tests directly as 'kyua-cli' under the tests hierarchy
Change the installation directory of tests from testsdir/kyua/kyua-cli/
to simply testsdir/kyua-cli/.  This goes towards the spirit of having one
subdirectory per package in testsdir.

This also avoids the need to have to install a Kyuafile in testsdir/kyua/.
I'm not sure which component should do this when we later have multiple
packages for Kyua, so it's just easier if we do not have to do it.

This change was r150 in Subversion.