Use ATF_REQUIRE to check for a NULL pointer instead of ATF_REQUIRE_EQ
If one of the arguments to ATF_REQUIRE_EQ is a NULL pointer, gcc may raise
a warning saying that NULL is being passed to operator<<.  This breaks the
build in NetBSD due to our use of -Werror.  Change this code to use a
check with ATF_REQUIRE instead.

This change was r152 in Subversion.