Make 'make check' work before doing a 'make install'
Separate the environment of the 'check' and 'installcheck' targets so
that the environment for 'check' can point to not-yet-installed files
and thus users can validate the new build before installing it.

This is a bit fragile: if the user does a './kyua test' before install,
some tests will still fail.  But this should not be done by hand anyway
so it is better to have 'make check' work.

This change was r154 in Subversion.