Convert the --show flag of the about command into an argument
Controlling what the about command shows with a flag (--show) instead of
an argument is confusing: it behaves in a different way than help, yet the
two commands are pretty similar.  Also, --show=all and no flag make things
more complex than necessary because there are two UIs that expose the same
behavior and these two interfaces need independent testing.

Change the about command to take what to show as an optional argument
instead of through a --show flag.

This change was r165 in Subversion.