Define a test_program interface and provide an ATF implementation
The previous test_program module was a simple collection of functions but
provided no "identity" to test programs.  This change adds a test_program
abstract base class and an atf_test_program implementation, which will
later allow us to provide different implementations if desired.

Note that, at the moment, this only generalizes the listing of test cases.
The generalization of the execution will come later when we do the same
kind of splitting in the test_case class.

This change was r166 in Subversion.