Define a test_case interface and provide an ATF implementation
The previous test_case module was a simple implementation of an ATF-based
test case but provided no genericity.  This change adds a test_case
abstract base class and an atf_test_case implementation, which will later
allow us to provide different implementations if desired.

This now generalizes the execution of the test cases through the new
virtual run() method.

As a result of this change, it's clear that the structure of runner and
results is tied to ATF and that these two modules need to be refactored.
The runner module is completely ATF-specific and should remain as such.
However, the results module probably deserves a similar abstract/specific
splitting as both test_program and test_case.

Indeed, this change is quite messy.  This is just an artifact of the poor
abstractions in the existing code and things will be cleaned up afterwards.

This change was r168 in Subversion.