Properly report test case setup failures
Instead of just crashing the test case setup routine (due to an unhandled
exception) when there is a problem, properly report the breakage message
as the test case result.

This is helpful in diagnosing problems with the execution of test cases
because, instead of getting an useless "signal received" message, we get
a descriptive message of the failure.

In particular, this helped me diagnose some breakage I was experiencing
in my system in which some unprivileged tests failed to execute because
the binary was unaccessible to the unprivileged user.

This change was r169 in Subversion.