Move ATF-specific routines from the results module into the new atf_results
The results module included both the definition of test results and provided
access to ATF-specific functions and logic.  Following the splitting of
test programs and test cases, this change starts doing the same for the
results handling.

The splitting is still incomplete as there is a fundamental problem in the
results module: the results classes currently defined in the module model
the particular semantics of the ATF test cases.  For example, there probably
is no point in exposing an expected_signal result outside of the ATF code.
Changes to solve this will come separately.

This change was r170 in Subversion.