Split the representation of ATF test case results from the generic ones
This change adds a new class to atf_results, raw_result, that provides
a mechanism to represent and manage the results of ATF-based test cases.
This class is supposed to be used internally only, and as a intermediate
step before constructing the generic test case result classes found in
the results module.

At the moment, the results module remains unchanged.  It still contains
a representation for all possible result types generated by ATF.  However,
this is wrong because the types in the results module should be generic
across different test interfaces.  This issue will be addressed later.

This change was r172 in Subversion.