Move the atf-specific code into a new atf_iface subdirectory
This change moves all of the code required to specialize the test program
and test case classes for atf into a new atf_iface subdirectory.  This
separates implementation from definition, and will allow to easily add
different interface implementations later on.

As part of this change, rename the abstract test_case and test_program
classes to base_test_case and base_test_program respectively.  This is to
prevent naming confusions.

This change was r173 in Subversion.