Implement the plain test interface
Plain test programs are programs that do not use any framework and report
their pass/fail status as an exit code.  Support for these is added to
simplify the integration of legacy test programs into a test suite, and
also to demonstrate that the run-time engine is generic enough to support
different test interfaces.

Note that this code is quite rudimentary and preliminary.  In particular:

1) There is a lot of code duplication with the atf interface implementation.
2) There is no support to tune some run-time features of the plain test
   programs; for example, to define a timeout.

These shortcomings will be addressed separately as they will require some
intrusive changes in unrelated parts of the tree.

This change was r179 in Subversion.